src/requirements.txt
opencv-python>=4.10.0
mediapipe>=0.10.14
numpy>=1.26.0
PyYAML>=6.0.1
Pillow>=10.0.0
